Bug correct - FN (FP) related to pos (neg) misclassification rate
The positive misclassification rate as used in the code of the shiny app and as explained in your paper is related to the misclassified samples among those with positive ground truth. Those are the false-negative samples. Same for negative MC rate. Therefore, at the 2 positions in the code FN and FP should be flipped to obtain correct results. 
Otherwise great app and very relevant paper!